requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv  # If you're using environment variables from a .env file
pandas  # Optional, for data processing and analysis
matplotlib  # Optional, for data visualization
